Add testing around actions and triggers.
